==> 05_calc_event_properties/P06_compile_azimuth_results.ipynb <==
## Compile the results of azimuth determination into one catalog

### Import libraries

import pandas as pd
from datetime import datetime, timedelta
import numpy as np
import os
import glob
import sys

### Load high-quality moonquake catalog

# Load
mqdir1 = '../catalogs/final_catalogs/'
cat = pd.read_csv(mqdir1 + 'A17_moonquakes_catalog_HQ_final.csv')
cat.drop(list(cat.filter(regex='Unnamed|index|mean_cc')), axis=1, inplace=True)

# Identify Grade AA and BB events
cat_AA_BB = cat.loc[(cat.grade_new == 'AA') | (cat.grade_new == 'BB')]
evids_hq = np.unique(cat_AA_BB.evid.tolist())

### Iteratively compile azimuth results

#### Lag

azdir = '/data/ytamama/Apollo17/catalogs/cc_azimuths/'
evids = []
thetas = []
xs_fin = []
ys_fin = []
misfits_fin = []

for evid in evids_hq:
    fname = glob.glob(f'{azdir}*{evid}_azimuth_results_lag.csv')[0]
    cat = pd.read_csv(fname)
    evids = np.concatenate([evids,cat.evid.tolist()])
    thetas = np.concatenate([thetas,cat.theta.tolist()])
    xs_fin = np.concatenate([xs_fin,cat.xs_fin.tolist()])
    ys_fin = np.concatenate([ys_fin,cat.ys_fin.tolist()])
    misfits_fin = np.concatenate([misfits_fin,cat.misfit_fin.tolist()])
    

d = {'evid':evids, 'theta_deg':thetas, 'xs_fin':xs_fin, 'ys_fin':ys_fin,
    'misfits_fin':misfits_fin}
lag_df = pd.DataFrame(data = d)
mqdir = '../catalogs/azimuths/'
lag_df.to_csv(mqdir + 'A17_moonquakes_AA_BB_azimuths_lag.csv',index=False)

#### SNR

evids = []
thetas = []
xs_fin = []
ys_fin = []
misfits_fin = []

for evid in evids_hq:
    fname = glob.glob(f'{azdir}*{evid}_azimuth_results_SNR.csv')[0]
    cat = pd.read_csv(fname)
    evids = np.concatenate([evids,cat.evid.tolist()])
    thetas = np.concatenate([thetas,cat.theta.tolist()])
    xs_fin = np.concatenate([xs_fin,cat.xs_fin.tolist()])
    ys_fin = np.concatenate([ys_fin,cat.ys_fin.tolist()])
    misfits_fin = np.concatenate([misfits_fin,cat.misfit_fin.tolist()])
    

d = {'evid':evids, 'theta_deg':thetas, 'xs_fin':xs_fin, 'ys_fin':ys_fin,
    'misfits_fin':misfits_fin}
snr_df = pd.DataFrame(data = d)
snr_df.to_csv(mqdir + 'A17_moonquakes_AA_BB_azimuths_SNR.csv',index=False)
