==> 05_calc_event_properties/P09_separate_isolated_repeating.ipynb <==
## Separate catalogs into isolated vs. repeating events

### Import libraries

import pandas as pd
import numpy as np
import os
import glob
import sys

### Load catalogs

# Catalog with information by geophone
mqdir1 = '../catalogs/final_catalogs/'
cat1 = pd.read_csv(mqdir1 + 'A17_moonquakes_catalog_HQ_final.csv')
cat1.drop(list(cat1.filter(regex='Unnamed|index')), axis=1, inplace=True)

# Catalog with information averaged per event
cat2 = pd.read_csv(mqdir1 + 'A17_moonquakes_catalog_HQ_avg_event_stats.csv')
cat2.drop(list(cat2.filter(regex='Unnamed|index')), axis=1, inplace=True)

### Separate isolated events

isolated_cat1 = cat1.loc[cat1.isol_or_rpt == 'isolated']
isolated_cat1 = isolated_cat1.reset_index()
isolated_cat1.drop(list(isolated_cat1.filter(regex='Unnamed|index')), axis=1, inplace=True)
mqdir2 = mqdir1 + 'isolated_vs_repeating/'
isolated_cat1.to_csv(mqdir2 + 'A17_isolated_catalog_HQ_final.csv',index=False)

isolated_cat2 = cat2.loc[cat2.isol_or_rpt == 'isolated']
isolated_cat2 = isolated_cat2.reset_index()
isolated_cat2.drop(list(isolated_cat2.filter(regex='Unnamed|index')), axis=1, inplace=True)
isolated_cat2.to_csv(mqdir2 + 'A17_isolated_catalog_HQ_avg_event_stats.csv',index=False)

len(isolated_cat2)

### Separate repeating events

repeating_cat1 = cat1.loc[cat1.isol_or_rpt == 'repeating']
repeating_cat1 = repeating_cat1.reset_index()
repeating_cat1.drop(list(repeating_cat1.filter(regex='Unnamed|index')), axis=1, inplace=True)
repeating_cat1.to_csv(mqdir2 + 'A17_repeating_catalog_HQ_final.csv',index=False)

repeating_cat2 = cat2.loc[cat2.isol_or_rpt == 'repeating']
repeating_cat2 = repeating_cat2.reset_index()
repeating_cat2.drop(list(repeating_cat2.filter(regex='Unnamed|index')), axis=1, inplace=True)
repeating_cat2.to_csv(mqdir2 + 'A17_repeating_catalog_HQ_avg_event_stats.csv',index=False)

len(repeating_cat2)
